refactored angualr app - removed unused imports - removed exports from authentication module. This should be a routing module only. - renamed authentication components and put them in seperate folders - removed barrel module from authentication. Don't need a barrel module here. - removed spec files. This in only a demo and we are not testing it
