+++ src/AngularClient/ClientApp/src/app/app.component.html
@@ -1,4 +1,1 @@
-<h1>
-  {{title}}
-</h1>
 <router-outlet></router-outlet>
