+++ src/AngularClient/ClientApp/src/app/authentication/authentication-routing.module.ts
@@ -1,9 +1,9 @@
-﻿import { NgModule } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { SigninCallbackComponent } from './components/signin/signin-callback.component';
-import { SignoutCallbackComponent } from './components/signout/signout-callback.component';
-import { SignedoutComponent } from './components/signout/signedout.component';
+import { SigninCallbackComponent } from './components/signin-callback/signin-callback.component';
+import { SignoutCallbackComponent } from './components/signout-callback/signout-callback.component';
+import { SignedoutComponent } from './components/signedout/signedout.component';
 
 const routes: Routes = [
     {
