+++ src/AngularClient/ClientApp/src/app/authentication/components/signedout/signedout.component.html
@@ -1,0 +1,3 @@
+<p>
+  You are now signed out.
+</p>
