--- src/AngularClient/ClientApp/src/app/authentication/components/signin/signin.component.html
@@ -1,1 +1,0 @@
-﻿<button (click)="signin()">Sign In</button>
