--- src/AngularClient/ClientApp/src/app/authentication/components/signout/signedout.component.html
@@ -1,3 +1,0 @@
-﻿<p>
-  You are now signed out. <signin></signin>
-</p>
