--- src/AngularClient/ClientApp/src/app/authentication/components/signout/signout.component.html
@@ -1,1 +1,0 @@
-﻿<button (click)="signout()">Sign Out</button>
