--- src/AngularClient/ClientApp/src/app/authentication/index.ts
@@ -1,1 +1,0 @@
-﻿export * from './authentication.module';
