+++ src/AngularClient/ClientApp/src/app/home/home.component.html
@@ -1,1 +1,1 @@
-﻿<signout></signout>
+<button (click)="signout()">Sign Out</button>
