+++ src/AngularClient/ClientApp/src/app/home/home.component.ts
@@ -1,15 +1,17 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
+import { AuthService } from 'libs/shared';
 
 @Component({
-  selector: 'boardroom-home',
+  selector: 'app-home',
   templateUrl: './home.component.html',
   styleUrls: ['./home.component.scss']
 })
-export class HomeComponent implements OnInit {
+export class HomeComponent {
 
-  constructor() { }
+  constructor(private authService: AuthService) { }
 
-  ngOnInit() {
+  signout() {
+    this.authService.signout();
   }
 
 }
